# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  internal_api.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: secure_password
        type: string
    required:
    - email
    - password
    type: object
  internal_api.LoginResponse:
    properties:
      refresh_token:
        example: ""
        type: string
      token:
        example: eyJhbGciOiJS...
        type: string
      user:
        properties:
          username:
            example: john.doe@example.com
            type: string
        type: object
    type: object
  internal_api.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Verve API
  version: "1.0"
paths:
  /auth/link/oauth:
    post:
      consumes:
      - application/json
      description: Link an OAuth account to the current user's account
      parameters:
      - description: OAuth provider (google/okta)
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Link OAuth account
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/unlink/oauth:
    post:
      description: Remove OAuth provider from user account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unlink OAuth account
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"